Move batch related code into a separate source file.
This removes a large amount of code from tiling.rs. The other
parts of tiling.rs can be moved to more appropriate files as
a follow up.
